requirements.txt
aniso8601==9.0.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
clickclick==20.10.2
colorama==0.4.4
Flask==2.1.1
Flask-RESTful==0.3.9
flask-restplus==0.13.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.0.0
wincertstore==0.2
zipp==3.8.0
